Moved web.js to adapters folder.
